requirements.txt
langchain-core
langchain-google-genai
langchain-community
python-dotenv